fix: whole-tree review findings, and release 0.7.7
A review pass over src/ and docs/ against real captures and the live
call sites. The findings that ship here:

HTTP: an HTTP source starts again after stop_streaming (the reader task
was spawned once at construction, so a SoapySDR deactivate/activate
cycle failed); the parser skips a packet whose body cannot be decoded
instead of failing on it forever while its buffer grew; HttpSource
resets the parser and drop detector on every reconnect; read_samples
returns a partial block on timeout instead of losing it; IqPacket
carries the rate the device streams, not the one requested; the
licence probe writes reflevel0 to the discovered receiver block, so it
can report Active; HttpSource refuses to build outside a Tokio runtime
rather than panicking in init; HttpSink with buffer_size 0 no longer
spins.

RTSA files: mSampleSize counts values per sample, not bytes (the
captures under tests/ carry 2 for float32 IQ and 1024 for spectra), so
seeking within a chunk and spectra reads used the wrong stride; reads
are bounded by the chunk's declared size; a negative DSFT stream offset
no longer overflows the header search; seek_to_sample works on
reverse-order files.

Native SDK: shutdown is refcounted so the last client, not the first,
tears the SDK down; a bare spectranv6eco opens spectranv6eco/rtsa;
read_samples_timeout honours its deadline; spectra reads honour the
packet stride; SDK calls that return no object are errors; opening
over a live device is refused; the device is closed on drop.

Also: clippy on Rust 1.98, README pinning a major behind the crate,
the C error-message entry point taking an int, seify reporting the
observed rate, and the cumulative_drops and bandwidth documentation.

+++ tests/c_api_test.rs
@@ -63,7 +63,7 @@ fn test_c_api_endpoints_client_lifecycle() {
 #[test]
 fn test_aaronia_get_error_message() {
     unsafe {
-        let msg_ptr = aaronia_get_error_message(AaroniaFfiError::NullPointer);
+        let msg_ptr = aaronia_get_error_message(AaroniaFfiError::NullPointer as i32);
         assert!(!msg_ptr.is_null());
         let msg = CStr::from_ptr(msg_ptr).to_string_lossy();
         assert_eq!(msg, "Null pointer provided");
